moved tests back to test_draw_utils - maybe they were moved in main after creating this branch?
